createme/updateme: streamline auto BSP creation
The existing automatic BSP creation was spread over two phases
of working on a git repository and hence the hand off was tricky
leading to improper branch usage in some situations.

With this change, all generation is done as part of updating the
feature description and can leverage the additional information
that is passed into update me to describe the machine being
built.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>