kern-tools: remove meta tag and directory assumptions
During repository sanity checks (createme) and during the
checkpoint process, there were several assumptions about the tree
that either relied on a tag, or a particular directory name.

With this set of changes, simply passing the meta branch name is
enough to sanitize and restore the checkpoint. If no meta branch
name is passed, the default of 'meta' is used for both the branch
and meta data directory name.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>