createme: test for non-zero meta_branch
If an explicit meta branch is not passed, createme should not test for
a directory of the same name to use as the meta directory.

In the scenario where a custom kernel repository is used and no meta
branch name is passed the meta branch name is empty. Testing for a
directory of "" always returns true, and the meta directory name
is set to "" as well.

This results in meta data writes to /cfg, /cfg/kernel-cache, etc, which
result in obvious permission denied errors.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>