updateme: anchor KMACHINE test
Out of tree feature descriptions (.scc files) take two forms: normal
features and BSP descriptions.

A normal feature is detected and added to the end of the current machine
being processed. During tree processing, it's configuration and patches
will be applied.

A BSP description on the other hand must be matched based on three
critera (which are in the .scc file via "define <foo>"):

  - machine
  - kernel type
  - architecture

Since features that define machines are only explicitly added, they
are removed from the list of features that should be automatically
added.

The criteria for removing them from the auto-add list is the
definitions found in the .scc file. The existing check was simply
for KMACHINE anywhere in the file. This meant that a conditional
or even a comment containing that phrase would exclude a file.

Properly anchoring the KMACHINE test to "^define.*KMACHINE" fixes the
problem of overly agreesive exclusions.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>