configme: derive path to tools from $0
If we are working on the tools, we want the ones being used to
be treated as a set.  So, if we are explicitly calling a configme
from the tool repo, we want merge_config.sh and other tools
to also be used from the repo, and not from whatever version
was captured in-tree on the meta branch.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>