scc: allow includes within conditional statements
A construct such as:

  if [ "$KARCH" = "x86" ]; then
      include feature.scc
  fi

wouldn't work, since the variable KARCH was not defined within
the limited environment used to determine the files that make
up a feature series.

Switching the detection of the files list from a subshell that
processes commands, to a simple grep check for 'include' solves
the problem, and simplfies the code.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>