scc: remove depreciated/unused commands
scc has gathered some obselete commands over the years. The
ability to indicate that patches should or shouldn't be included
based on certain conditions was supported when scc operated
outside of a git repository. The git based support for this ended
up being complex and error prone. It is no longer used in any
known feature descriptions, so the support can be removed from
scc.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>