kconf_check: simplify cmdline args, dont store data per branch
Largely due to historical reasons, there used to be a distinction
between the top of a "project" and the location of the kernel tree.
However, we currently were just passing in /dev/null to pad the args
and doing a pointless "if [ -d ... ]" on it.

Since the current instance ends up having `pwd` for both LINUX_SRC
and TOP_BUILD_DIR, just replace all instances of the latter with
the former.

With respect to the /dev/null, replace it with the $META file name.
We want this in kconf_check, since it was storing the audit data
on a per branch basis, and that is no longer unique now that several
BSPs re-use the same branch, resulting in overwritten audit data
when multiple builds were done in a single tree (i.e. buildall).
The same change is made in configme to KTGT; it was set to the
branch, and used in the path of audit data, but it is not unique.

As we aren't using the branch name at all anymore, drop it from
the args entirely.   This allows deleting the get_branch function,
and we assign KTGT=$KTYPE/$KMACHINE which should hopefully be
unique per build.

For BSPs that have their own unique branch, the above is consistent
with the current kgit-meta path creation naming behaviour, i.e.:

   linux$ls meta/cfg/standard/emenlow/
   ls: cannot access meta/cfg/standard/emenlow/: No such file or directory
   linux$kgit-meta -v -k meta/cfg/meta/emenlow-standard-meta
   [INFO] collecting configs in ./meta/meta-series
     [##################################################]  (completed in 1 seconds)
   linux$ls meta/cfg/standard/emenlow/
   config_frag.txt  hdw_frags.txt  non_hdw_frags.txt
   linux$

We see a slightly different behaviour from kgit-meta though,
when the branch is shared.  It had similar code that, on _kconf
processing, would use the non-unique branch name for placement
of audit data.  It also has been updated to use instead a KTGT
that is set from $KTYPE/$KMACHINE

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>