buildall: remove all instances of it using/reading scc files
The whole point of an scc file is to be the input that
facilitates the generation of the metadata that describes
a BSP branch (its patches, config etc).

So, once a meta-series exists, going back to the scc file to
retrieve data doesn't make any sense.  It is like ld going
back to look for something in a C file.

As we map $SCC --> $META, make both the META_DIR and the
META_BRANCH configurable, since there will be a need for
that going forward.  Current defaults are "meta" for both.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>