configme: catch errors found during fragment sanitization
During the pre merge processing of fragments and error is thrown
if a fragment doesn't exist. configme wasn't checking the return
code from the sanity script, and hence would continue processing
and produce a .config that would be missing options for any fragment
that followed the one in error.

Catching the error, and propagating it up to the calling build
system exposes this to the user .. who can then fix the problem.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>