scc: improve include processing
If an include file didn't exist, reporting of the error was
previously delated until the meta series was executed. This
is later than ideal, and can potentially be time consuming.

Moving the check into the file processing time, improves the
timing of the reporting.

There was also latent code for calling 'parent' routines,
but this is no longer used, so it is removed as part of this
improvement.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>