kgit-meta: prep the starting branch for patches
In the case of a custom repository, we may not enounter ANY explicit
branch points. This means that the first patch push will fail, since
the series file and other infrastructure has not been set up. To
ensure that we can always push patches, the starting branch is now
prepped for modifications.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>