kgit-meta: exclude explicit branches from name calculations
kernel branches are constructed during patching of the tree by
constructing a '/' based hierarchy of names as each branch
directive is encountered.

But if a "branch $name $branchpoint" is used, the entire branch
name is supplied so no additions to the hierarchy should
happen. As such, that type of branch command should not be part
of branch name calculation and preparation.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>