kconf_check: add "verify" configuration fragment type
To allow a BSP to verify that critical options have made it to the
final .config (and to throw an error if they are not present), a
new "verify" kernel config fragment type is introduced.

The format of this fragment is identical to any other fragment,
except that its contents will not be used to directly configure
the kernel, but will be checked against the final .config any
missing or changed options will result in a message and error exit
code.

This allows the BSP to enforce critical options, but not have to
override policy or declare non-hardware options as hardware.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>