updateme: find the board description with the highest score
When searching for the best matching board description the scripts
know how many defines are passed in. The .scc files are then searched
for the one which matches that definition.

The input defines were previously used to calculate both a minimum
score, which was then compared against the maximum score from a
given .scc file. This meant that the best match was not found if
only a single define matched. Having the ability to find the
description that has the highest score, but doesn't match all the
defines is useful when working with repositories with custom
meta data, since things like kernel type are not always defined
or relevant.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>