patching: only validate user supplied patches by default
Issue: LIN6-14197

Previously the patching tools would consider both system and user
supplied patches in the same manner .. they are simply a series of
patches to be applied to a branch, and that the scripts should determine
where in the series to start (based on what is already on the branch).

This detection was causing a few problems:

  - time consuming
  - starting in the middle of a series when intermediate patches were
    merged to a branch.

To solve both the performance and start detection, we instead simply
note the transition from system (i.e. already defined features and
series) and user/recipe supplied patches. When the transition is noted,
the system will start pushing ALL patches without doing autoresume
detection.

Control in keeping the series up to date is passed to the user, and
consistent behaviour/performance is achieved.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>