meta: protect branch counts from empty strings
On some machines (and kernel trees), it has been found that the branch
count may resolve to an empty string.

With the current patch/branch tests, this causes a patch application
error due to comparisons resolving with no right or left hand side.

To protect against this, we can combine the conditions and protect the
counts with quotes. This doesn't resolve the root cause, but is a
better, defensive condition.

Signed-off-by: Marek Vasut <marex@denx.de>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>