tools: handle directories with, or without, trailing /
Issue: LIN8-3482

As part of processing configuration, patches and other operations the
paths of the files are normalized so they can be relocated from where
they were originally processed.

This is done by recording the directories that contain patches and
configuration, and then removing the longest possible path from each
element. Once that is done, we can search by their unique suffix across
any number of target directory structures.

The code that generated the path names, and the code that searches for
the longest prefix do not always agree on whether a trailing slash is
present. To cover all cases, we can simply try once with the slash
and once without.

When the relocation code misses the longest element, we end up with
arcane processing errors later in the tools since configs and patches
are not where they are expected.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>