kgit-meta: always checkout branches on branch commands
During a tree generation we must always make the branch active when we
see any kind of branch command. This ensures that any subsequent patches
are applied in the proper context.

Previously, only branch creation was changing the active branch, and
this mean that tree generation was not determinstic and relied on the
order of processing to generate a correct tree.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>